fix(tests): flytt frist-aar-helpere til fristsjekk for CI uten nicegui
CI installerer kun [dev] og ikke [ui], så tester som importerer fra
wenche.ui feiler med ModuleNotFoundError: No module named 'nicegui'.

Helperne neste_frist() og regnskapsaar_for_frist() er rene dato-
funksjoner uten UI-avhengighet, så de hører naturlig hjemme i
wenche.fristsjekk. Flytter dem dit, sletter test_ui_helpers.py og
legger testene inn i test_fristsjekk.py i stedet.
